evaluation_metrics/pyproject.toml
[tool.poetry]
name = "evaluation-metrics"
version = "0.1.0"
description = ""
authors = ["jgosaliya <jgosaliya@dobledevops.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.37.1"
requests = "^2.32.3"
markdown = "^3.6"
beautifulsoup4 = "^4.12.3"
python-dotenv = "^1.0.1"
numpy = "^2.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
